Removed utils.py and model.py and put them into task.py as I think this makes more sense. Also some minor changes on filenames to avoid confusion.
Added function in server_app.py that is able to aggregate fit metrics (i.e. accuracy and validation), which is then sent to the strategy. This strategy saves that data to a local .json file and then uploads to wandb (similar to the other functions in the strategy).

Also started working some more on visualization (although not done yet), as well as the differential privacy part. Currently it is not working properly as including DP results in an abysmal aggregated accuracy.